+++ internal/server/config.go
@@ -2,14 +2,18 @@ package server
 
 import (
 	"log"
+	"os"
 	"time"
+
+	"devops-tpl/internal/storage/file"
+	"devops-tpl/internal/storage/memory"
 
 	"github.com/caarlos0/env/v6"
 )
 
 const (
 	SERVERADDRPORT = "localhost:8080"
-	STOREINTERVAL  = 0
+	STOREINTERVAL  = 10
 	STOREFILE      = "devops-metrics-db.json"
 	RESTORE        = true
 )
@@ -34,4 +38,63 @@ func GetEnvConfig() Config {
 	}
 	return cfg
 
+}
+
+func InitConfig() {
+	cfg := GetEnvConfig()
+	if cfg.RestoreSavedData {
+		reader, err := file.NewReader(cfg.StoreFile)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		checkFile, err := os.Stat(cfg.StoreFile)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		size := checkFile.Size()
+
+		if size == 0 {
+			writer, err := file.NewWriter(cfg.StoreFile)
+			if err != nil {
+				log.Fatal(err)
+			}
+
+			if err := writer.WriteDatabase(); err != nil {
+				log.Fatal(err)
+			}
+		}
+
+		if file.DB, err = reader.ReadDatabase(); err != nil {
+			log.Fatal(err)
+		}
+	}
+
+	if cfg.StoreFile != " " {
+		if cfg.StoreInterval == 0 {
+			for range memory.MetricUpdated {
+				writer, err := file.NewWriter(cfg.StoreFile)
+				if err != nil {
+					log.Fatal(err)
+				}
+
+				if err := writer.WriteDatabase(); err != nil {
+					log.Fatal(err)
+				}
+			}
+		} else {
+			ticker := time.NewTicker(cfg.StoreInterval)
+			for range ticker.C {
+				writer, err := file.NewWriter(cfg.StoreFile)
+				if err != nil {
+					log.Fatal(err)
+				}
+
+				if err := writer.WriteDatabase(); err != nil {
+					log.Fatal(err)
+				}
+			}
+		}
+	}
 }
